change the code, still uncomplete

+++ Backend/src/main/java/com/Maxeeey/TODOListElements/ITODOListElement.java
@@ -3,6 +3,6 @@ package com.Maxeeey.TODOListElements;
 public interface ITODOListElement {
 	public String getName();
 	public boolean getIsDone();
-	public int getId();
+	public Long getId();
 	public void changeIsDoneBoolean();
 }
